pyproject.toml
[tool.poetry]
name = "boraboat"
version = "0.1.0"
description = ""
authors = ["sininto <sininto@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
asgiref = "^3.8.1"
certifi = "^2024.2.2"
charset-normalizer = "^3.3.2"
colorama = "^0.4.6"
coreapi = "^2.3.3"
coreschema = "^0.0.4"
django-rest-swagger = "^2.2.0"
djangorestframework = "^3.15.1"
idna = "^3.7"
inflection = "^0.5.1"
itypes = "^1.2.0"
jinja2 = "^3.1.4"
markupsafe = "^2.1.5"
openapi-codec = "^1.3.2"
pillow = "^10.3.0"
pypng = "^0.20220715.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
qrcode = "^7.4.2"
requests = "^2.32.2"
simplejson = "^3.19.2"
sqlparse = "^0.5.0"
typing-extensions = "^4.11.0"
tzdata = "^2024.1"
uritemplate = "^4.1.1"
urllib3 = "^2.2.1"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.29.0"
gunicorn = "^22.0.0"
dj-database-url = "^2.1.0"
weasyprint = "^62.1"
reportlab = "^4.2.0"
chardet = "^5.2.0"
fpdf  = "^1.7.2"


[build-system]
requires = ["poetry-boraboat"]
build-backend = "poetry.boraboat.masonry.api"
